requirements.txt
rasa==3.6.20
rasa-sdk==3.6.2
mysql-connector-python==8.2.0
python-dotenv==1.0.0
uvicorn==0.24.0
fastapi==0.104.1